registration module some few setups
