+++ webapp/gradle.properties
@@ -1,0 +1,3 @@
+git_properties_plugin_version=2.2.2
+apt_plugin_version=0.21
+checkstyle_version=8.32
