--- webapp/src/main/java/cm/ubuea/covider/samplemodule/domain/SampleDomain.java
@@ -1,5 +1,0 @@
-package cm.ubuea.covider.samplemodule.domain;
-
-public class SampleDomain {
-    
-}
