--- webapp/src/main/java/cm/ubuea/covider/samplemodule/repository/SampleRepository.java
@@ -1,5 +1,0 @@
-package cm.ubuea.covider.samplemodule.repository;
-
-public class SampleRepository {
-    
-}
