--- webapp/src/main/java/cm/ubuea/covider/samplemodule/service/SampleService.java
@@ -1,5 +1,0 @@
-package cm.ubuea.covider.samplemodule.service;
-
-public class SampleService {
-    
-}
