+++ webapp/src/main/resources/application.properties
@@ -1,2 +1,9 @@
 spring.output.ansi.enabled=ALWAYS
+spring.banner.location=classpath:banner.txt
 spring.main.allow-bean-definition-overriding=true
+
+jwt.header=Authorization
+jwt.secret=covider
+jwt.expiration=604800
+jwt.route.authentication.path=/auth
+jwt.route.authentication.refresh=/refresh
